# Conflicts:
#	src/Nethermind/Nethermind.Optimism/InitializeBlockchainOptimism.cs

+++ src/Nethermind/Nethermind.TxPool/Collections/PersistentBlobTxDistinctSortedPool.cs
@@ -88,10 +88,20 @@ public class PersistentBlobTxDistinctSortedPool : BlobTxDistinctSortedPool
         return false;
     }
 
-    protected override bool Remove(ValueHash256 hash, Transaction tx)
+    protected override bool Remove(ValueHash256 hash, out Transaction? tx)
     {
-        _blobTxCache.Delete(hash);
-        _blobTxStorage.Delete(hash, tx.Timestamp);
-        return base.Remove(hash, tx);
+        if (base.Remove(hash, out tx))
+        {
+            if (tx is not null)
+            {
+                _blobTxStorage.Delete(hash, tx.Timestamp);
+            }
+
+            _blobTxCache.Delete(hash);
+
+            return true;
+        }
+
+        return false;
     }
 }
